python/requirements.txt
more_itertools==9.1.0
deprecation==2.1.0
parameterized==0.9.0
django-colorfield===0.8.0
voluptuous==0.13.1
django-test-migrations==1.3.0
mockito==1.4.0
kink==0.6.6
cachetools==5.3.0
psycopg2==2.9.6
supabase~=1.0.0
asgiref==3.6.0
coverage==7.2.5
dj-database-url==2.0.0
Django~=4.2.1
django-cors-headers==3.14.0
django-extensions==3.2.1
gunicorn==20.1.0
Pillow==9.5.0
pytz==2023.3
selenium==4.9.0
sqlparse==0.4.4
urllib3==2.0.2
whitenoise==6.4.0
python-dotenv==1.0.0